refactor(scanner): extract cli/scanner_report.py — fmt + save_log + main + get_top_symbols (#225 PR7)
